0. >>> ht = {}

+++ 0x1A-hash_tables/0-hash_table_create.c
@@ -8,7 +8,7 @@
 */
 hash_table_t *hash_table_create(unsigned long int size)
 {
-	unsigned int iterator;
+	unsigned long int iterator;
 	hash_table_t *ht = malloc(sizeof(hash_table_t));
 
 	if (ht == NULL || size <= 0)
